fixed decimal for int long

+++ printf_integer.c
@@ -7,12 +7,12 @@
  */
 int _printf_integer(va_list args)
 {
-	int num = va_arg(args, int);
+	long int num = va_arg(args, long int);
 
 	/* this is the entire extract from the variadic argument list */
 
 	int count = 0; /* count: it is counter the number print */
-	int temp; /* it is a variable temp for stock the num */
+	long int temp; /* it is a variable temp for stock the num */
 	char buffer[50]; /* it is array containing the digits of the integer */
 	int index = 0;
